+++ CHANGELOG.md
@@ -29,11 +29,18 @@
 
 ## API Changes
 
+* Add `POST /conversations/list-ids` (#1686)
+* Deprecate `GET /converstations/ids` (#1686)
+
 ## Features
 
 * Client functions for the hscim library (#1699)
 
 ## Bug fixes and other updates
+
+## Federation changes (alpha feature, do not use yet)
+
+* Add new API to list paginated qualified conversation ids (#1686)
 
 ## Documentation